requirements.txt
streamlit
pandas
feedparser
python-dotenv
openai
newspaper3k
plotly
wordcloud
matplotlib
azure-storage-blob
BeautifulSoup
chardet